Relocate packages under root package `pybatsim`
PyBatsim shipped two packages `batsim` and `schedulers` as top-level
packages.

This commits is a first step towards a neater architecture for pybatsim.

`batsim` is a misleading name as it let users think this is a pure
Python implementation of Batsim: this is not the purpose of this
package.

`schedulers` is easy to be confused with the standard library `sched`
module.