Implement detection of ambiguous schedulers
The names of entry points is not guaranteed to be unique across all
packages: schedulers defined in different packages may use the same
entry point name.
If a user requests to run a scheduler (identified by the name of its
entry point) that is bound to multiple classes, the execution is aborted
as there is no way to univocally choose a scheduler.
Note that the execution continues as long as the user requests to run a
scheduler bound to a single class, even if some other ambiguous
schedulers exist.