nix: use a single /default.nix
This will make simpler the importation of this git repository from Nix code,
as a common hypothesis is that a default.nix is present at its root.

Other nix files could have been kept in each subfolder,
but as their Nix description is small I feel like it will be simpler to maintain
by only using a single file.