Fix output of pybatsim --list-external-decision-components
According to the documentation, inspect.getdoc() behavior is to
"retrieve the documentation string from the inheritance hierarchy".

As we now derive from typing.Protocol, if a user does not write a
docstring for their class, we inherit the docstring of typing.Protocol
in the description of the EDCs.

This commit drops this behavior by directly using __doc__, and only
displays the docstring first non-empty line of the user's class.
Should the docstring not exist, we fall back to the __qualname__ of the
user's class.