Update existing job in place for dynamic job ack
Dynamic jobs are created by the EDC.
They are stored in _alive_jobs when the EDC creates a RegisterJobEvent.
Should the EDC request the acknowledgment of dynamic jobs, Batsim will
send back the job description.
This commit introduces the required deserialization logic to reuse the
existing job.
This is required as the pybatsim API works with Job objects.